refactor(server): extract metering boot into boot-metering.ts (#4243 step 3)
